Refaktorer routes
- Fjerner duplikatkode med en layout-route
- Legger all logikken i en egen routes-fil og bruker index-fila for eksportering
